+++ src/store/emotionStore.ts
