+++ src/types/Emotions.ts
@@ -1,0 +1,7 @@
+export type Emotions = {
+  id?: number;
+  user_id: number;
+  data: string;
+  emocao: string;
+  descricao: string;
+};
