+++ src/types/User.ts
@@ -1,0 +1,6 @@
+export type User = {
+  id?: number;
+  nome: string;
+  email: string;
+  password: string;
+};
